ENGR00132249 USB:Fix dead loop when cat iomem
  Apply this fix for mx25.  Has been seperately cherry-picked for
other platforms.

  The origin code reuse the otg_resource as the resource for udc
device. This will cause a small bug, the following command will
print endless log:
	#cat /proc/iomem

  The reason of the bug is that reuse the otg_resources for the
udc device will register the otg_resources twice which will make
the ->child and ->parent of the otg_resources pointer to himself.
So the above command will traverse in a dead loop for ever.

Signed-off-by: Huang Shijie <b32955@freescale.com>