ENGR00132153 sgtl5000: do not enforce symmetry if rate is unknown
If a playback and a record stream are started together, the
alsa userspace may not not do the necessary ioctl calls for
the audio driver stack to know the sample rate and sample
size of the first stream before starting the second stream,

i.e.:
  arecord -D plughw:0 -d 10 -f S16_LE -r 44100 -c 2 -traw |
  aplay -D plughw:0 -f S16_LE -r 44100 -c 2 -traw

Normally we want to the driver to make sure that the second
stream does not request a different playback rate/bit size
from the first stream.  But at that point, the driver thinks
that the sample rate and sample size are 0.  So in that case
do not enforce constraints.

This still allows for usecases where play is started first and
then record is started after a delay while playback continues.

Signed-off-by: Alan Tull <alan.tull@freescale.com>